Create bugsite
This was the Product Landing Page certification project for freeCodeCamp. I like the way it looks, but I will gladly take any feedback on my code or how I could make this page better. 

I am new to GitHub so if I posted this in the wrong place please inform me.